refactor: move distribution related stuff to the separate package
