+++ slider/src/main/java/io/monstarlab/mosaic/slider/SliderState.kt
@@ -11,6 +11,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.geometry.Offset
+import io.monstarlab.mosaic.slider.distribution.SliderValueDistribution
 import kotlinx.coroutines.coroutineScope
 
 /**
