+++ slider/src/main/java/io/monstarlab/mosaic/slider/distribution/LinearEquation.kt
@@ -1,4 +1,4 @@
-package io.monstarlab.mosaic.slider
+package io.monstarlab.mosaic.slider.distribution
 
 public data class LinearEquation(
     private val m: Float,
