+++ slider/src/main/java/io/monstarlab/mosaic/slider/distribution/SliderValueDistribution.kt
@@ -1,4 +1,4 @@
-package io.monstarlab.mosaic.slider
+package io.monstarlab.mosaic.slider.distribution
 
 import androidx.annotation.FloatRange
 import kotlin.math.sqrt
@@ -57,46 +57,5 @@ public interface SliderValueDistribution {
                 return value
             }
         }
-    }
-}
-
-internal fun SliderValueDistribution.interpolate(
-    range: ClosedFloatingPointRange<Float>,
-): ClosedFloatingPointRange<Float> {
-    return interpolate(range.start)..interpolate(range.endInclusive)
-}
-
-internal fun SliderValueDistribution.inverse(
-    range: ClosedFloatingPointRange<Float>,
-): ClosedFloatingPointRange<Float> {
-    return inverse(range.start)..inverse(range.endInclusive)
-}
-
-/**
- * Represents a parabolic distribution strategy for slider values.
- *
- * This strategy calculates interpolated and inversely interpolated values based on a parabolic equation:
- * f(x) = a * x^2 + b * x + c
- *
- * @property a coefficient of the x^2 term in the parabolic equation
- * @property b coefficient of the x term in the parabolic equation
- * @property c constant term in the parabolic equation
- * @constructor Creates a ParabolicValueDistribution with the given coefficients.
- */
-public class ParabolicValueDistribution(
-    private val a: Float,
-    private val b: Float,
-    private val c: Float,
-) : SliderValueDistribution {
-
-    override fun inverse(value: Float): Float {
-        if (value == 0f) return 0f
-
-        val d = (b * b) - 4 * a * (c - value)
-        return (-b + sqrt(d)) / (2 * a)
-    }
-
-    override fun interpolate(value: Float): Float {
-        return a * (value * value) + b * value + c
     }
 }
